requirements.txt
backports.functools-lru-cache==1.4
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
decorator==4.1.2
funcsigs==1.0.2
h5py==2.7.1
html5lib==0.9999999
Keras==2.0.8
mahotas==1.4.3
Markdown==2.2.0
matplotlib==2.1.0
mock==2.0.0
networkx==2.0
numpy==1.13.3
olefile==0.44
palettable==3.1.0
pbr==3.1.1
Pillow==4.3.0
protobuf==3.4.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
sklearn==0.0
subprocess32==3.2.7
tensorflow-gpu==1.2.0
tensorflow-tensorboard==0.1.8
Werkzeug==0.12.2
